backend/requirements.txt
# Python Environment (usa Python 3.11+)
# python=3.11

# FastAPI & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Email
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Utilities
python-dateutil>=2.8.2
email-validator>=2.1.0
